+++ src/main/kotlin/ml/duncte123/skybot/web/controllers/api/GetUserGuilds.kt
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package ml.duncte123.skybot.web.routes.api
+package ml.duncte123.skybot.web.controllers.api
 
 import com.jagrosh.jdautilities.oauth2.OAuth2Client
 import com.jagrosh.jdautilities.oauth2.entities.OAuth2Guild
