+++ src/main/kotlin/ml/duncte123/skybot/web/controllers/api/Kpop.kt
@@ -16,39 +16,34 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package ml.duncte123.skybot.web.routes.api
+package ml.duncte123.skybot.web.controllers.api
 
 import ml.duncte123.skybot.Author
 import ml.duncte123.skybot.connections.database.DBManager
 import ml.duncte123.skybot.utils.ApiUtils
-import net.dv8tion.jda.bot.sharding.ShardManager
 import org.json.JSONObject
+import spark.Request
 import spark.Response
+import java.sql.SQLException
 
 @Author(nickname = "duncte123", author = "Duncan Sterken")
-object MainApi {
+object Kpop {
 
-    fun serverCount(response: Response, shardManager: ShardManager): Any {
-        return JSONObject()
-            .put("status", "success")
-            .put("server_count", shardManager.guildCache.size())
-            .put("shard_count", shardManager.shardsTotal)
-            .put("code", response.status())
-    }
+    fun show(request: Request, response: Response, database: DBManager): Any {
+        val search = request.queryParamOrDefault("search", "")
 
-    fun joinGuild(response: Response) {
-        response.redirect("https://discord.gg/NKM9Xtk")
-    }
+        return try {
+            ApiUtils.getRandomKpopMember(database, search).toJson()
+                .put("status", "success")
+                .put("code", response.status())
 
-    fun llama(response: Response, database: DBManager): Any {
-        return ApiUtils.getRandomLlama(database).toJson()
-            .put("status", "success")
-            .put("code", response.status())
-    }
+        } catch (e: SQLException) {
+            response.status(404)
 
-    fun alpaca(response: Response): Any {
-        return ApiUtils.getRandomAlpaca().toJson()
-            .put("status", "success")
-            .put("code", response.status())
+            JSONObject()
+                .put("status", "faiure")
+                .put("message", "Nothing found")
+                .put("code", response.status())
+        }
     }
 }
